refactor: 파일 디렉터리 변경 (#55)
