+++ backend/src/test/java/ku_rum/backend/domain/user/dto/request/EmailValidationRequestTest.java
@@ -2,7 +2,7 @@ package ku_rum.backend.domain.user.dto.request;
 
 import jakarta.validation.Validation;
 import jakarta.validation.Validator;
-import ku_rum.backend.domain.user.dto.request.mail.EmailValidationRequest;
+import ku_rum.backend.domain.mail.dto.request.EmailValidationRequest;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
